Tick backend/requirements.txt
admin-tools==0.1
anyio==4.1.0
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
chart-tools==0.1.3
chartjs==1.2
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Cython==3.0.5
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-csp==3.7
django-debug-toolbar==4.2.0
django-el-pagination==4.0.0
django-extensions==3.2.3
django-grappelli==3.0.8
django-import-export==3.3.5
django-jazzmin==2.6.0
django-jet==1.0.8
django-js-asset==2.1.0
django-memoize==2.3.1
django-model-tracker==2.0.4
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-role-permissions==3.2.0
django-settings==1.3.12
django-social-share==2.3.0
django-sslserver==0.22
django-tables2==2.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
fonttools==4.41.1
gunicorn==21.2.0
h11==0.14.0
hijri==0.1.2
hijri-converter==2.3.1
httpcore==1.0.2
httpx==0.25.2
idna==3.4
imageio==2.31.3
import-export==0.3.1
Jinja2==3.1.2
jsonfield==3.1.0
jsonfield2==4.0.0.post0
kiwisolver==1.4.4
lml==0.1.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
primer==1.4.4
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-nvd3==0.14.2
pytz==2023.3
raven==6.10.0
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
seaborn==0.12.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sr==0.1.0
tablib==3.5.0
text-unidecode==1.3
texttable==1.7.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.4
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
